Добавлен прогноз погоды по координатам,а не только в г.Санкт-Петербург

+++ app/src/main/java/com/myapp/weather/dto/daily/Wind.kt
@@ -1,6 +1,6 @@
 package com.myapp.weather.dto.daily
 
 data class Wind(
-    val speed: Int? = null,
+    val speed: Double? = null,
     val deg: Int? = null
 )
